# api_interface_demo/swagger_server/models/create_sensor.py
# coding: utf-8

from __future__ import absolute_import
from datetime import date, datetime  # noqa: F401

from typing import List, Dict  # noqa: F401

from swagger_server.models.base_model_ import Model
from swagger_server import util


class CreateSensor(Model):
    """NOTE: This class is auto generated by the swagger code generator program.

    Do not edit the class manually.
    """
    def __init__(self, name: str=None, sensor_type: str=None, device_id: str=None):  # noqa: E501
        """CreateSensor - a model defined in Swagger

        :param name: The name of this CreateSensor.  # noqa: E501
        :type name: str
        :param sensor_type: The sensor_type of this CreateSensor.  # noqa: E501
        :type sensor_type: str
        :param device_id: The device_id of this CreateSensor.  # noqa: E501
        :type device_id: str
        """
        self.swagger_types = {
            'name': str,
            'sensor_type': str,
            'device_id': str
        }

        self.attribute_map = {
            'name': 'name',
            'sensor_type': 'sensorType',
            'device_id': 'device_id'
        }
        self._name = name
        self._sensor_type = sensor_type
        self._device_id = device_id

    @classmethod
    def from_dict(cls, dikt) -> 'CreateSensor':
        """Returns the dict as a model

        :param dikt: A dict.
        :type: dict
        :return: The CreateSensor of this CreateSensor.  # noqa: E501
        :rtype: CreateSensor
        """
        return util.deserialize_model(dikt, cls)

    @property
    def name(self) -> str:
        """Gets the name of this CreateSensor.


        :return: The name of this CreateSensor.
        :rtype: str
        """
        return self._name

    @name.setter
    def name(self, name: str):
        """Sets the name of this CreateSensor.


        :param name: The name of this CreateSensor.
        :type name: str
        """

        self._name = name

    @property
    def sensor_type(self) -> str:
        """Gets the sensor_type of this CreateSensor.


        :return: The sensor_type of this CreateSensor.
        :rtype: str
        """
        return self._sensor_type

    @sensor_type.setter
    def sensor_type(self, sensor_type: str):
        """Sets the sensor_type of this CreateSensor.


        :param sensor_type: The sensor_type of this CreateSensor.
        :type sensor_type: str
        """
        allowed_values = ["SENSOR_NODE", "ASSET_SENSOR"]  # noqa: E501
        if sensor_type not in allowed_values:
            raise ValueError(
                "Invalid value for `sensor_type` ({0}), must be one of {1}"
                .format(sensor_type, allowed_values)
            )

        self._sensor_type = sensor_type

    @property
    def device_id(self) -> str:
        """Gets the device_id of this CreateSensor.


        :return: The device_id of this CreateSensor.
        :rtype: str
        """
        return self._device_id

    @device_id.setter
    def device_id(self, device_id: str):
        """Sets the device_id of this CreateSensor.


        :param device_id: The device_id of this CreateSensor.
        :type device_id: str
        """

        self._device_id = device_id